refactor(components): rename camel case files to pascal case
