+++ src/components/DeleteButton.svelte
